+++ tests/CRUDlexTests/CRUDServiceProviderTest.php
@@ -231,4 +231,13 @@ class CRUDServiceProviderTest extends \PHPUnit_Framework_TestCase {
         $this->assertSame($read, $expected);
     }
 
+    public function testGetLanguageName() {
+        $crudServiceProvider = new CRUDServiceProvider();
+        $expected = 'Deutsch';
+        $read = $crudServiceProvider->getLanguageName('de');
+        $this->assertSame($read, $expected);
+        $read = $crudServiceProvider->getLanguageName('invalid');
+        $this->assertNull($read);
+    }
+
 }